v 1.19 修改dht11采集失败bug

+++ main/bsp/dht11/bsp_dht11.h
@@ -9,5 +9,5 @@
 #define DHT11_TASK_STACK 4096
 
 esp_err_t bsp_dht11_init(void);
-esp_err_t bsp_dht11_read(float *humidity, float *temperature);
+esp_err_t bsp_dht11_read(int *humidity, float *temperature);
 void      bsp_dht11_task(void *arg);
